feat: redesign landing page + fix phantom BigInt error
Landing page:
- Hero: product preview browser mockup with live tool cards
- Features: bento grid layout with glow hover effects and API code preview
- Testimonials: infinite dual-row marquee (pause on hover)
- CTA: full-bleed dark section with perks checklist

Phantom fix:
- Pass BigInt to createTransferInstruction (spl-token v0.4.x requirement)
- Use Number() consistently in server-side balance verification

+++ src/components/billing/phantom-button.tsx
@@ -64,7 +64,7 @@ export function PhantomButton({ plan, className, size = "sm", variant = "outline
       const treasuryKey = new PublicKey(treasuryAddress);
       const usdcMint = new PublicKey(USDC_MINT);
 
-      const rawAmount = PLAN_USDC[plan] * 10 ** USDC_DECIMALS;
+      const rawAmount = BigInt(PLAN_USDC[plan] * 10 ** USDC_DECIMALS);
 
       const senderATA = await getAssociatedTokenAddress(usdcMint, senderKey);
       const receiverATA = await getAssociatedTokenAddress(usdcMint, treasuryKey);
